+++ src/app/src/main/kotlin/org/imperial/mrc/hint/security/SecurityConfig.kt
@@ -14,6 +14,7 @@ import org.pac4j.core.profile.ProfileManager
 import org.pac4j.core.util.Pac4jConstants
 import org.pac4j.http.client.indirect.FormClient
 import org.pac4j.jee.context.session.JEESessionStore
+import org.pac4j.oauth.config.OAuth20Configuration.STATE_REQUEST_PARAMETER
 import org.pac4j.sql.profile.service.DbProfileService
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.ComponentScan
@@ -22,7 +23,6 @@ import org.springframework.jdbc.datasource.TransactionAwareDataSourceProxy
 import org.springframework.stereotype.Component
 import java.util.*
 import javax.sql.DataSource
-
 
 @Configuration
 @ComponentScan(basePackages = ["org.pac4j.springframework.web", "org.pac4j.springframework.component"])
@@ -61,6 +61,14 @@ class Session(
     {
         private const val VERSION_ID = "version_id"
         private const val MODE = "mode"
+    }
+
+    fun generateStateParameter(): String
+    {
+        val state = UUID.randomUUID().toString()
+        val encodedState = Base64.getEncoder().encodeToString(state.toByteArray())
+        sessionStore.set(webContext, STATE_REQUEST_PARAMETER, encodedState)
+        return encodedState
     }
 
     fun getUserProfile(): CommonProfile