+++ src/app/src/main/kotlin/org/imperial/mrc/hint/security/oauth2/OAuth2AuthenticationRedirection.kt
@@ -1,6 +1,7 @@
 package org.imperial.mrc.hint.security.oauth2
 
 import org.imperial.mrc.hint.AppProperties
+import org.imperial.mrc.hint.security.Session
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -11,7 +12,7 @@ import java.net.URI
 @Component
 class OAuth2AuthenticationRedirection(
     protected val appProperties: AppProperties,
-    protected val oauth2State: OAuth2State
+    protected val session: Session? = null,
 )
 {
     fun oauth2LoginRedirect(): ResponseEntity<String>
@@ -21,7 +22,7 @@ class OAuth2AuthenticationRedirection(
             .path("/authorize")
             .queryParam("response_type", "code")
             .queryParam("client_id", appProperties.oauth2ClientId)
-            .queryParam("state", oauth2State.generateCode())
+            .queryParam("state", session?.generateStateParameter())
             .queryParam("scope", "openid+profile+email+read:dataset")
             .queryParam("audience", appProperties.oauth2ClientAudience)
             .queryParam("redirect_uri", "${appProperties.applicationUrl}/callback/oauth2Client")