+++ NEWS.md
@@ -1,3 +1,7 @@
+# hint.2.19.0
+
+* Verify auth0 state for protection against threat of cross-site request forgery 
+
 # hint.18.0
 
 * Update naomi ADR integration to pull 2023 datasets